services/ai-server/requirements.txt
# Core API Framework
fastapi==0.116.1
uvicorn[standard]==0.35.0

# Data Validation
pydantic==2.11.7
pydantic-settings==2.3.4

# AI and Machine Learning
transformers==4.55.4
torch==2.8.0
accelerate==1.10.0
llama-cpp-python==0.3.6

# HTTP Client and File Handling
httpx==0.27.2
aiofiles==23.2.1
python-multipart==0.0.20

# Environment and Configuration
python-dotenv==1.1.1

# Image Processing (for vision tasks)
Pillow==11.3.0

# System utilities
psutil==7.0.0
